requirements.txt
beautifulsoup4
requests
telegram==0.0.1
python-telegram-bot==13.7
emoji
python-dotenv